requirements.txt
streamlit
pandas
altair
playwright
trafilatura
beautifulsoup4
python-dotenv
